pyproject.toml
[project]
name = "pdf-to-audio"
version = "0.1.0"
description = "Convert German PDF files to audio using Piper TTS"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "piper-tts>=1.2.0",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
]

[project.scripts]
pdf-to-audio = "pdf_to_audio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
